Systematic pass fixing diverging functions on the gcc 2.7.2.3 ELF oracle (gladi386.so) to byte-exact match, verified never to regress the MSVC6 PE oracle (gladiator.dll). Full ELF (audit_so.py) and PE (audit_dll.sh) audits both confirm 0 regressions against the git-tracked baselines.
Newly confirmed full byte-identical matches on BOTH oracles this batch:

  pattern — flattened `if (ms->jumpreach) {...}` into an early-return guard
  instead of an indented wrapper block, matching gcc's block layout.
- BotTravel_Teleport (be_ai_move.c): removed a redundant `v4 = ms->moveflags`
  local; real re-reads `ms->moveflags` from memory independently at each use
- BotFinishTravel_WeaponJump (be_ai_move.c): same cross-jump flattening as
  BotFinishTravel_Jump, plus reordering `hordir[2]=0.0f` to follow (not
- BotMoveInGoalArea (be_ai_move.c): reverse-declaration-order stack layout
  fix (moveresult declared before dir).
- LoadItemConfig (be_ai_goal.c): reverse-declaration-order fix across three
  reorder (cfg->items assignment before cfg->numitems=0). Documented that
  token_t is 0x42C bytes under gcc vs 0x430 under MSVC (l_script.h,
  long double field alignment difference).
- BotInitLevelItems (be_ai_goal.c): fixed statement order (ic=itemconfig
  declaration-order change could not fix — the modelindex-fixup loop counter
  and the later item-name search loop counter are actually ONE reused source
  a single `i`).

Also includes accumulated verified fixes from prior sessions across
be_aas_bspq2.c, be_aas_cluster.c, be_aas_entity.c, be_aas_file.c,
be_ea.c, l_libvar.c, l_precomp.c, l_script.c, l_struct.c — all part of the
same ELF-oracle sweep, all covered by the 0-regression full-audit
funcmap_overrides.tsv updated to reflect current function mapping;
(AAS_IndexFromString loop-guard shape).

Native (gcc, aarch64) and MinGW (i686-w64-mingw32-gcc) cross-builds both
verified clean after this batch.

Source-commit: 8f6fb2580c03d73fc0cd1cf81061604e6aa9e90d

+++ botlib/be_ai2_dmq2.h
@@ -64,7 +64,7 @@ void __cdecl BotCTFRetreatGoals(bot_state_t *bs);
 void __cdecl BotCTFSeekGoals(bot_state_t *bs);
 int __cdecl BotCTFTeam(bot_state_t *bs);
 BOOL BotCanAndWantsToRocketJump(bot_state_t *bs);
-double __cdecl BotChatTime(bot_state_t *bs);
+float __cdecl BotChatTime(bot_state_t *bs);
 int __cdecl BotChat_Death(int *bs);
 int __cdecl BotChat_EndLevel(bot_state_t *bs);
 BOOL __cdecl BotChat_EnterGame(bot_state_t *bs);
